LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/nested-sharedworker-success.https-expected.txt

CONSOLE MESSAGE: line 7: ReferenceError: Can't find variable: SharedWorker

Harness Error (FAIL), message = ReferenceError: Can't find variable: SharedWorker
